use source files instead facets
