+++ lib/FilterNavigation/FilterNavigation.js
@@ -12,10 +12,7 @@ import { navigationSegments } from '../constants';
 const FilterNavigation = () => {
   const {
     navigationSegmentValue,
-    setSearchDropdownValue,
-    setSearchIndex,
     setNavigationSegmentValue,
-    defaultDropdownValueBySegment,
   } = useContext(AuthoritiesSearchContext);
 
   return (
@@ -32,13 +29,7 @@ const FilterNavigation = () => {
             role="tab"
             id={`segment-navigation-${name}`}
             data-testid={`segment-navigation-${name}`}
-            onClick={() => {
-              const value = defaultDropdownValueBySegment[name] || '';
-
-              setSearchDropdownValue(value);
-              setSearchIndex(value);
-              setNavigationSegmentValue(name);
-            }}
+            onClick={() => setNavigationSegmentValue(name)}
           >
             <FormattedMessage id={`stripes-authority-components.label.${name}`} />
           </Button>
