+++ lib/constants/filterConfig/filterConfig.js
@@ -1,15 +1,6 @@
+import { buildDateRangeQuery } from '../../utils';
 import { AUTH_REF_TYPES } from '../authRefTypes';
 import { REFERENCES_VALUES_MAP } from '../references';
-
-const buildDateRangeQuery = name => values => {
-  const [startDateString, endDateString] = values[0]?.split(':') || [];
-
-  if (!startDateString || !endDateString) {
-    return '';
-  }
-
-  return `(metadata.${name}>="${startDateString}" and metadata.${name}<="${endDateString}")`;
-};
 
 export const FILTERS = {
   CREATED_DATE: 'createdDate',
