+++ lib/constants/filterConfig/filterConfig.test.js
@@ -4,26 +4,24 @@ import {
   FILTERS,
 } from './filterConfig';
 
+jest.mock('../../utils', () => ({
+  buildDateRangeQuery: () => jest.fn().mockReturnValue('called buildDateRangeQuery'),
+}));
+
 describe('filterConfig', () => {
   describe('createdDate filter', () => {
     it('should call buildDateRangeQuery', () => {
-      const values = ['2022-11-01:2022-12-07'];
-      const result = '(metadata.createdDate>="2022-11-01" and metadata.createdDate<="2022-12-07")';
-
       const filter = filterConfig.find(config => config.name === 'createdDate');
 
-      expect(filter.parse(values)).toEqual(result);
+      expect(filter.parse()).toEqual('called buildDateRangeQuery');
     });
   });
 
   describe('updatedDate filter', () => {
     it('should call buildDateRangeQuery', () => {
-      const values = ['2022-11-01:2022-12-07'];
-      const result = '(metadata.updatedDate>="2022-11-01" and metadata.updatedDate<="2022-12-07")';
-
       const filter = filterConfig.find(config => config.name === 'updatedDate');
 
-      expect(filter.parse(values)).toEqual(result);
+      expect(filter.parse()).toEqual('called buildDateRangeQuery');
     });
   });
