+++ lib/context/AuthoritiesSearchContext.js
@@ -12,6 +12,7 @@ import { buildFiltersObj } from '@folio/stripes-acq-components';
 import { useDidUpdate } from '../hooks';
 import {
   navigationSegments,
+  searchableIndexesValues,
   FILTERS,
 } from '../constants';
 
@@ -20,16 +21,21 @@ const propTypes = {
     PropTypes.arrayOf(PropTypes.node),
     PropTypes.node,
   ]).isRequired,
-  defaultDropdownValueBySegment: PropTypes.object,
+  initialValues: PropTypes.object,
   readParamsFromUrl: PropTypes.bool,
 };
 
 const AuthoritiesSearchContext = createContext();
 
+const defaultDropdownValueBySegment = {
+  [navigationSegments.search]: searchableIndexesValues.KEYWORD,
+  [navigationSegments.browse]: '',
+};
+
 const AuthoritiesSearchContextProvider = ({
   children,
   readParamsFromUrl,
-  defaultDropdownValueBySegment,
+  initialValues,
 }) => {
   const location = useLocation();
   const filterUrlParams = Object.values(FILTERS);
@@ -39,18 +45,19 @@ const AuthoritiesSearchContextProvider = ({
     : {};
 
   const getInitialFilters = () => {
-    return pick(buildFiltersObj(queryString.stringify(searchParams)), filterUrlParams);
+    return initialValues.filters || pick(buildFiltersObj(queryString.stringify(searchParams)), filterUrlParams);
   };
 
   const initialSegment = searchParams.segment || navigationSegments.search;
 
-  const initialDropdownValue = searchParams.qindex || defaultDropdownValueBySegment[initialSegment] || '';
+  const initialDropdownValue = searchParams.qindex || initialValues.dropdownValue || defaultDropdownValueBySegment[initialSegment];
+  const initialSearchIndex = searchParams.qindex || (initialValues.searchIndex ?? defaultDropdownValueBySegment[initialSegment]);
 
-  const [navigationSegmentValue, setNavigationSegmentValue] = useState(initialSegment); // eslint-disable-line react/hook-use-state
+  const [navigationSegmentValue, _setNavigationSegmentValue] = useState(initialSegment); // eslint-disable-line react/hook-use-state
   const [searchInputValue, setSearchInputValue] = useState(searchParams.query || '');
   const [searchQuery, setSearchQuery] = useState(searchParams.query || '');
   const [searchDropdownValue, setSearchDropdownValue] = useState(initialDropdownValue);
-  const [searchIndex, setSearchIndex] = useState(initialDropdownValue);
+  const [searchIndex, setSearchIndex] = useState(initialSearchIndex);
   const [filters, setFilters] = useState(getInitialFilters());
   const [advancedSearchRows, setAdvancedSearchRows] = useState([]);
   const [isGoingToBaseURL, setIsGoingToBaseURL] = useState(false);
@@ -59,12 +66,18 @@ const AuthoritiesSearchContextProvider = ({
     option: searchParams.qindex,
   });
 
+  const setNavigationSegmentValue = value => {
+    setSearchDropdownValue(initialValues.dropdownValue || defaultDropdownValueBySegment[value]);
+    setSearchIndex(initialValues.searchIndex ?? defaultDropdownValueBySegment[value]);
+    _setNavigationSegmentValue(value);
+  };
+
   const resetAll = () => {
     setSearchInputValue('');
     setSearchQuery('');
-    setSearchDropdownValue(defaultDropdownValueBySegment[navigationSegmentValue] || '');
-    setSearchIndex(defaultDropdownValueBySegment[navigationSegmentValue] || '');
-    setFilters({});
+    setSearchDropdownValue(initialValues.dropdownValue || defaultDropdownValueBySegment[navigationSegmentValue]);
+    setSearchIndex(initialValues.searchIndex ?? defaultDropdownValueBySegment[navigationSegmentValue]);
+    setFilters(initialValues.filters || {});
     setIsGoingToBaseURL(true);
     setAdvancedSearchDefaultSearch(null);
   };
@@ -93,7 +106,6 @@ const AuthoritiesSearchContextProvider = ({
     setIsGoingToBaseURL,
     advancedSearchDefaultSearch,
     setAdvancedSearchDefaultSearch,
-    defaultDropdownValueBySegment,
   };
 
   return (
@@ -107,7 +119,7 @@ const AuthoritiesSearchContextProvider = ({
 
 AuthoritiesSearchContextProvider.propTypes = propTypes;
 AuthoritiesSearchContextProvider.defaultProps = {
-  defaultDropdownValueBySegment: {},
+  initialValues: {},
   readParamsFromUrl: true,
 };
