+++ lib/queries/useAuthorities/useAuthorities.js
@@ -5,16 +5,65 @@ import {
 } from 'react';
 import { useQuery } from 'react-query';
 import queryString from 'query-string';
+import template from 'lodash/template';
 
 import {
   useOkapiKy,
   useNamespace,
 } from '@folio/stripes/core';
+import { defaultAdvancedSearchQueryBuilder } from '@folio/stripes/components';
 
-import { QUERY_KEY_AUTHORITIES } from '../../constants';
-import { getCqlQueryForSearchLookup } from '../../utils';
+import { buildQuery } from '../utils';
+import {
+  filterConfig,
+  searchableIndexesValues,
+  QUERY_KEY_AUTHORITIES,
+} from '../../constants';
 
 const AUTHORITIES_API = 'search/authorities';
+
+const buildRegularSearch = (searchIndex, query, filters) => {
+  const compileQuery = template(
+    buildQuery({
+      searchIndex,
+      filters,
+      query,
+    }),
+    { interpolate: /%{([\s\S]+?)}/g },
+  );
+
+  const cqlSearch = [];
+
+  if (query) {
+    const queryParam = searchIndex === searchableIndexesValues.IDENTIFIER
+      ? query
+      : query.trim();
+
+    const compiledQuery = compileQuery({ query: queryParam });
+
+    cqlSearch.push(compiledQuery);
+  }
+
+  return cqlSearch;
+};
+
+const buildAdvancedSearch = (advancedSearch, filters) => {
+  const rowFormatter = (index, query, comparator) => {
+    const compileQuery = template(
+      buildQuery({
+        searchIndex: index,
+        comparator,
+        filters,
+        query,
+      }),
+      { interpolate: /%{([\s\S]+?)}/g },
+    );
+
+    return compileQuery({ query });
+  };
+
+  return [defaultAdvancedSearchQueryBuilder(advancedSearch, rowFormatter)];
+};
 
 const useAuthorities = ({
   searchQuery,
@@ -25,7 +74,6 @@ const useAuthorities = ({
   pageSize,
   sortOrder,
   sortedColumn,
-  enabled,
 }) => {
   const ky = useOkapiKy();
   const [namespace] = useNamespace({ key: QUERY_KEY_AUTHORITIES });
@@ -43,15 +91,27 @@ const useAuthorities = ({
     sortedColumn,
   ]);
 
-  const cqlQuery = getCqlQueryForSearchLookup({
-    isAdvancedSearch,
-    advancedSearch,
-    filters,
-    searchIndex,
-    searchQuery,
-    sortOrder,
-    sortedColumn,
-  });
+  let cqlSearch = [];
+
+  if (isAdvancedSearch) {
+    cqlSearch = buildAdvancedSearch(advancedSearch, filters);
+  } else {
+    cqlSearch = buildRegularSearch(searchIndex, searchQuery, filters);
+  }
+
+  const cqlFilters = Object.entries(filters)
+    .filter(([, filterValues]) => filterValues.length)
+    .map(([filterName, filterValues]) => {
+      const filterData = filterConfig.find(filter => filter.name === filterName);
+
+      return filterData.parse(filterValues);
+    });
+
+  let cqlQuery = [...cqlSearch, ...cqlFilters].join(' and ');
+
+  if (sortOrder && sortedColumn) {
+    cqlQuery += ` sortBy ${sortedColumn}/sort.${sortOrder}`;
+  }
 
   const searchParams = {
     query: cqlQuery,
@@ -75,7 +135,6 @@ const useAuthorities = ({
       return ky.get(path).json();
     }, {
       keepPreviousData: true,
-      enabled,
     },
   );
