+++ lib/queries/useAuthoritiesBrowse/useAuthoritiesBrowse.js
@@ -13,11 +13,12 @@ import {
   useNamespace,
 } from '@folio/stripes/core';
 
-import { QUERY_KEY_AUTHORITIES } from '../../constants';
+import { buildHeadingTypeQuery } from '../utils';
 import {
-  getCqlQueryForBrowseLookup,
-  startingSearchForBrowseLookup,
-} from '../../utils';
+  filterConfig,
+  QUERY_KEY_AUTHORITIES,
+  searchableIndexesValues,
+} from '../../constants';
 
 const AUTHORITIES_BROWSE_API = 'browse/authorities';
 
@@ -28,17 +29,25 @@ const useBrowseRequest = ({
   startingSearch,
   pageSize,
   precedingRecordsCount,
-  enabled,
 }) => {
   const ky = useOkapiKy();
   const [namespace] = useNamespace({ key: QUERY_KEY_AUTHORITIES });
 
+  const cqlSearch = startingSearch ? [startingSearch] : [];
 
-  const query = getCqlQueryForBrowseLookup({
-    startingSearch,
-    searchIndex,
-    filters,
-  });
+  cqlSearch.push(`isTitleHeadingRef==${searchIndex === searchableIndexesValues.NAME_TITLE}`);
+
+  const cqlFilters = Object.entries(filters)
+    .filter(([, filterValues]) => filterValues.length)
+    .map(([filterName, filterValues]) => {
+      const filterData = filterConfig.find(filter => filter.name === filterName);
+
+      return filterData ? filterData.parse(filterValues) : null;
+    });
+
+  const headingTypeQuery = buildHeadingTypeQuery(searchIndex);
+
+  const query = [...cqlSearch, ...cqlFilters, headingTypeQuery].filter(Boolean).join(' and ');
 
   const searchParams = {
     query,
@@ -63,7 +72,6 @@ const useBrowseRequest = ({
     }, {
       keepPreviousData: true,
       staleTime: 5 * 60 * 1000,
-      enabled,
     },
   );
 
@@ -99,7 +107,7 @@ const useBrowserPaging = (initialQuery, initialIndex) => {
     }
 
     if (newPage === 0) {
-      newMainRequestSearch = startingSearchForBrowseLookup(initialQuery);
+      newMainRequestSearch = [`(headingRef>="${initialQuery}" or headingRef<"${initialQuery}")`];
     }
 
     return newMainRequestSearch;
@@ -145,7 +153,6 @@ const useAuthoritiesBrowse = ({
   searchIndex,
   pageSize,
   precedingRecordsCount,
-  enabled,
 }) => {
   const [currentQuery, setCurrentQuery] = useState(searchQuery);
   const [currentIndex, setCurrentIndex] = useState(searchIndex);
@@ -171,7 +178,6 @@ const useAuthoritiesBrowse = ({
     searchIndex: currentIndex,
     pageSize,
     precedingRecordsCount,
-    enabled,
   });
 
   const prevPageRequest = useBrowseRequest({
@@ -181,7 +187,6 @@ const useAuthoritiesBrowse = ({
     searchIndex: currentIndex,
     pageSize,
     precedingRecordsCount,
-    enabled,
   });
 
   const nextPageRequest = useBrowseRequest({
@@ -191,7 +196,6 @@ const useAuthoritiesBrowse = ({
     searchIndex: currentIndex,
     pageSize,
     precedingRecordsCount,
-    enabled,
   });
 
   const allRequestsFetched = mainRequest.isFetched && prevPageRequest.isFetched && nextPageRequest.isFetched;
