+++ lib/queries/useFacets/useFacets.js
@@ -11,7 +11,6 @@ const FACETS_API = 'search/authorities/facets';
 const useFacets = ({
   query,
   selectedFacets,
-  enabled,
 }) => {
   const ky = useOkapiKy();
   const [namespace] = useNamespace({ key: 'useFacets' });
@@ -21,7 +20,7 @@ const useFacets = ({
     facet: selectedFacets.join(','),
   };
 
-  const { isFetching, data, refetch } = useQuery(
+  const { isFetching, data } = useQuery(
     [namespace, searchParams],
     async () => {
       const path = `${FACETS_API}?${queryString.stringify(searchParams)}`.replace(/\+/g, '%20');
@@ -29,15 +28,12 @@ const useFacets = ({
       return ky.get(path).json();
     }, {
       keepPreviousData: true,
-      enabled,
-      staleTime: 300,
     },
   );
 
   return ({
     facets: data?.facets,
     isLoading: isFetching,
-    refetch,
   });
 };
