+++ lib/utils.js
@@ -1,18 +1,16 @@
 import omit from 'lodash/omit';
 import cloneDeep from 'lodash/cloneDeep';
 import set from 'lodash/set';
-import template from 'lodash/template';
 
-import { defaultAdvancedSearchQueryBuilder } from '@folio/stripes/components';
+export const buildDateRangeQuery = name => values => {
+  const [startDateString, endDateString] = values[0]?.split(':') || [];
 
-import {
-  filterConfig,
-  searchableIndexesValues,
-} from './constants';
-import {
-  buildHeadingTypeQuery,
-  buildQuery,
-} from './queries/utils';
+  if (!startDateString || !endDateString) {
+    return '';
+  }
+
+  return `(metadata.${name}>="${startDateString}" and metadata.${name}<="${endDateString}")`;
+};
 
 export const getSelectedFacets = filterAccordions => {
   return Object.keys(filterAccordions).filter(accordion => filterAccordions[accordion]);
@@ -68,107 +66,4 @@ export const markHighlightedFields = (marcSource, authority) => {
   set(marcSourceClone, 'data.parsedRecord.content.fields', marcFields);
 
   return marcSourceClone;
-};
-
-const buildRegularSearch = (searchIndex, query, filters) => {
-  const compileQuery = template(
-    buildQuery({
-      searchIndex,
-      filters,
-      query,
-    }),
-    { interpolate: /%{([\s\S]+?)}/g },
-  );
-
-  const cqlSearch = [];
-
-  if (query) {
-    const queryParam = searchIndex === searchableIndexesValues.IDENTIFIER
-      ? query
-      : query.trim();
-
-    const compiledQuery = compileQuery({ query: queryParam });
-
-    cqlSearch.push(compiledQuery);
-  }
-
-  return cqlSearch;
-};
-
-const buildAdvancedSearch = (advancedSearch, filters) => {
-  const rowFormatter = (index, query, comparator) => {
-    const compileQuery = template(
-      buildQuery({
-        searchIndex: index,
-        comparator,
-        filters,
-        query,
-      }),
-      { interpolate: /%{([\s\S]+?)}/g },
-    );
-
-    return compileQuery({ query });
-  };
-
-  return [defaultAdvancedSearchQueryBuilder(advancedSearch, rowFormatter)];
-};
-
-export const getCqlQueryForSearchLookup = ({
-  isAdvancedSearch,
-  advancedSearch,
-  filters,
-  searchIndex,
-  searchQuery = '',
-  sortOrder,
-  sortedColumn,
-}) => {
-  let cqlSearch;
-
-  if (isAdvancedSearch) {
-    cqlSearch = buildAdvancedSearch(advancedSearch, filters);
-  } else {
-    cqlSearch = buildRegularSearch(searchIndex, searchQuery, filters);
-  }
-
-  const cqlFilters = Object.entries(filters)
-    .filter(([, filterValues]) => filterValues.length)
-    .map(([filterName, filterValues]) => {
-      const filterData = filterConfig.find(filter => filter.name === filterName);
-
-      return filterData.parse(filterValues);
-    });
-
-  let cqlQuery = [...cqlSearch, ...cqlFilters].join(' and ');
-
-  if (sortOrder && sortedColumn) {
-    cqlQuery += ` sortBy ${sortedColumn}/sort.${sortOrder}`;
-  }
-
-  return cqlQuery;
-};
-
-export const startingSearchForBrowseLookup = (initialQuery = '') => {
-  return [`(headingRef>="${initialQuery}" or headingRef<"${initialQuery}")`];
-};
-
-export const getCqlQueryForBrowseLookup = ({
-  startingSearch,
-  searchIndex,
-  filters,
-}) => {
-  const cqlSearch = startingSearch ? [startingSearch] : [];
-
-  cqlSearch.push(`isTitleHeadingRef==${searchIndex === searchableIndexesValues.NAME_TITLE}`);
-
-  const cqlFilters = Object.entries(filters)
-    .filter(([, filterValues]) => filterValues.length)
-    .map(([filterName, filterValues]) => {
-      const filterData = filterConfig.find(filter => filter.name === filterName);
-
-      return filterData ? filterData.parse(filterValues) : null;
-    });
-
-  const headingTypeQuery = buildHeadingTypeQuery(searchIndex);
-
-  return [...cqlSearch, ...cqlFilters, headingTypeQuery].filter(Boolean).join(' and ');
 };
